Feat&Bug: Implement timer in screen_game. BUT timebase messed up
In simulator, everything works as expected.
But when download hex to STM32F7, the timebase is faster than it should be.
In screen_game, the timer(widget: digiClock) updates too fast and slow down when screen is touched.
So you can't simply change timebase to fix this.

This can be fix by enableing freeRTOS.
The clock example of touchGFX works perfect both in simulator and STM32F7 with freeRTOS,
and has same bug on STM32F7 without freeRTOS.